refactor(State): 상태 패키지 분류
