+++ src/main/java/blackjack/participant/state/playing/PlayerState.java
@@ -1,7 +1,11 @@
-package blackjack.participant.state;
+package blackjack.participant.state.playing;
 
 import blackjack.card.Card;
 import blackjack.participant.Hand;
+import blackjack.participant.state.GameState;
+import blackjack.participant.state.terminated.BlackJackState;
+import blackjack.participant.state.terminated.BustedState;
+import blackjack.participant.state.terminated.StandState;
 
 public class PlayerState extends GameState {
