+++ src/test/java/blackjack/participant/state/GameStateTest.java
@@ -5,6 +5,10 @@ import static org.assertj.core.api.Assertions.assertThat;
 import blackjack.card.Card;
 import blackjack.card.Number;
 import blackjack.card.Shape;
+import blackjack.participant.state.playing.DealerState;
+import blackjack.participant.state.playing.PlayerState;
+import blackjack.participant.state.terminated.BlackJackState;
+import blackjack.participant.state.terminated.StandState;
 import java.util.List;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
